rules slice
