+++ src/hooks/useGeneratorStore/slices/index.ts
@@ -1,0 +1,2 @@
+export * from './loadProfile'
+export * from './rules'
